refactor: move cli trace helpers to `reth-tracing`
